refactor: remove unused config-persistence code (#16167)
